Add logic - only staff can CUD pet

+++ src/controllers/pets.ts
@@ -12,9 +12,21 @@ const getById = async (ctx: Context, next: any) => {
     await next()
 }
 
+const filter = async (ctx: Context, next: any) => {
+    const requestBody = <Pets>ctx.request.body
+    ctx.body = await petsRepository.filter(requestBody)
+    await next()
+}
+
 const add = async (ctx: Context, next: any) => {
     const requestBody = <Pets>ctx.request.body
-    requestBody.created_by = <Users>ctx.state.user.data
+    const currentUser = <Users>ctx.state.user.data
+    if(!currentUser.staff_code){
+        ctx.body = {message: 'User is not a staff!'}
+        ctx.status = 200
+        return
+    }
+    requestBody.created_by = currentUser
     ctx.body = await petsRepository.create(<Pets>ctx.request.body)
     ctx.status = 200
     await next()
@@ -22,7 +34,13 @@ const add = async (ctx: Context, next: any) => {
 
 const update = async (ctx: Context, next: any) => {
     const requestBody = <Pets>ctx.request.body
-    requestBody.updated_by = <Users>ctx.state.user.data
+    const currentUser = <Users>ctx.state.user.data
+    if(!currentUser.staff_code){
+        ctx.body = {message: 'User is not a staff!'}
+        ctx.status = 200
+        return
+    }
+    requestBody.updated_by = currentUser
     const result = await petsRepository.update(+ctx.params.id, <Pets>ctx.request.body)
     ctx.body = result ? result : {message: "Invalid request!"}
     ctx.status = 200
@@ -31,9 +49,14 @@ const update = async (ctx: Context, next: any) => {
 
 const remove = async (ctx: Context, next: any) => {
     const currentUser = <Users>ctx.state.user.data
+    if(!currentUser.staff_code){
+        ctx.body = {message: 'User is not a staff!'}
+        ctx.status = 200
+        return
+    }
     ctx.body = await petsRepository.remove(+ctx.params.id, currentUser) ? {message: "Item deleted successfully"} : {message: "Item not found!"}
     ctx.status = 200
     await next()
 }
 
-export default {getAll, getById, add, update, remove}
+export default {getAll, getById, filter, add, update, remove}
